Add nonsecure image loading handshake
Updated the loading of the nonsecure image to be more of a handshake between
the boot code and the nonsecure code.  The new approach allows the boot code to
simply jump to the nonsecure loader at image offset 0 (instead of 4).  The
nonsecure loader then hands back the nonsecure initialization entry point that
the boot code can supply to the secure initialization.

Signed-off-by: Greg Bellows <greg.bellows@linaro.org>